[Refactor] Dto 어노테이션 수정
필요없는 Dto 삭제 및 어노테이션 수정
