+++ src/main/java/com/nhnacademy/bookpubshop/category/dto/request/CreateCategoryRequestDto.java
@@ -1,7 +1,6 @@
 package com.nhnacademy.bookpubshop.category.dto.request;
 
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Length;
@@ -22,10 +21,8 @@ public class CreateCategoryRequestDto {
 
     private Integer parentCategoryNo;
 
-    @NotNull(message = "카테고리 우선순위는 필수 사항입니다.")
     private Integer categoryPriority;
 
-    @NotNull(message = "카테고리 노출여부는 필수 사항입니다.")
     private boolean categoryDisplayed;
 
 }
